Fix module import issues with Extract
